requirements.txt
discord.py==2.4.0
aiosqlite==0.19.0
python-dotenv==1.0.1
rapidfuzz==3.9.4
loguru==0.7.2
pytest==8.3.3
